store_monitoring/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
pandas==2.1.4
pytz==2023.3
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.115.0